# clients/go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: email@gmail.com
  description: |
    This is a sample server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Open APi
  version: 1.0.0
servers:
- url: http://localhost:8080
...
paths:
...
  /post:
    post:
      operationId: addPost
      requestBody:
        $ref: '#/components/requestBodies/Post'
      responses:
        "200":
          description: Post created successfully
        "405":
          description: Invalid input
      security:
      - cm_auth:
        - write
        - read
      summary: Add a new post
      tags:
      - post
    put:
      operationId: updatePost
      requestBody:
        $ref: '#/components/requestBodies/Post'
      responses:
        "200":
          description: Post updated successfully
        "400":
          description: Invalid ID supplied
        "404":
          description: Comment not found
        "405":
          description: Validation exception
      security:
      - cm_auth:
        - write
        - read
      summary: Update an existing post
      tags:
      - post
  /post/{postId}:
    delete:
      operationId: deletePost
      parameters:
      - explode: false
        in: header
        name: api_key
        required: false
        schema:
          type: string
        style: simple
      - description: Post id to delete
        explode: false
        in: path
        name: postId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Post deleted successfully
        "400":
          description: Invalid ID supplied
        "404":
          description: Post not found
      security:
      - cm_auth:
        - write
        - read
      summary: Deletes a post
      tags:
      - post
    get:
      description: Returns a single post
      operationId: getPostById
      parameters:
      - description: ID of post to return
        explode: false
        in: path
        name: postId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
            application/xml:
              schema:
                $ref: '#/components/schemas/Post'
          description: Post fetched successfully
        "400":
          description: Invalid ID supplied
        "404":
          description: Comment not found
      security:
      - api_key: []
      summary: Find post by ID
      tags:
      - post
    post:
      operationId: updatePostWithForm
      parameters:
      - description: ID of post that needs to be updated
        explode: false
        in: path
        name: postId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/updatePostWithForm_request'
      responses:
        "200":
          description: Post updated successfully
        "405":
          description: Invalid input
      security:
      - cm_auth:
        - write
        - read
      summary: Updates a post with form data
      tags:
      - post
...
  /user:
    post:
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: Created user object
        required: true
      responses:
        "200":
          description: User created successfully
      summary: Create user
      tags:
      - user
  /user/login:
    get:
      operationId: loginUser
      parameters:
      - description: The user name for login
        explode: true
        in: query
        name: username
        required: true
        schema:
          type: string
        style: form
      - description: The password for login in clear text
        explode: true
        in: query
        name: password
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
            application/xml:
              schema:
                type: string
          description: User login successfully
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              explode: false
              schema:
                format: int32
                type: integer
              style: simple
            X-Expires-After:
              description: date in UTC when token expires
              explode: false
              schema:
                format: date-time
                type: string
              style: simple
        "400":
          description: Invalid username/password supplied
      summary: Logs user into the system
      tags:
      - user
...
  /user/{username}:
    delete:
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
      - description: The name that needs to be deleted
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      responses:
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
      summary: Delete user
      tags:
      - user
    get:
      operationId: getUserByName
      parameters:
      - description: The name that needs to be fetched. Use user1 for testing.
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
          description: User fetched successfully
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
      summary: Get user by user name
      tags:
      - user
    put:
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
      - description: name that need to be updated
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        description: Updated user object
        required: true
      responses:
        "400":
          description: Invalid user supplied
        "404":
          description: User not found
      summary: Updated user
      tags:
      - user
components:
  requestBodies:
...
    Post:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Post'
        application/xml:
          schema:
            $ref: '#/components/schemas/Post'
      description: Post object that needs to be added
...
  schemas:
...
    Post:
      example:
        postType: postType
        id: id
        title: title
        excerpt: excerpt
        url: url
        content: content
        commentStatus: Disabled
      properties:
        id:
          type: string
        title:
          type: string
        url:
          type: string
        excerpt:
          type: string
        content:
          type: string
        commentStatus:
          enum:
          - Disabled
          - RequireReview
          - AutomateReview
          type: string
        postType:
          type: string
      type: object
      xml:
        name: Post
    User:
      example:
        password: password
        role:
          capabilities:
          - capabilities
          - capabilities
          name: name
        avatarUrl: avatarUrl
        verificationStatus:
          isVerified: true
          verificationMethod: verificationMethod
          verificationTimestamp: 2000-01-23T04:56:07.000+00:00
        displayName: displayName
        blogs:
        - authToken: authToken
          name: name
          id: id
          url: url
          commentStatus: Disabled
        - authToken: authToken
          name: name
          id: id
          url: url
          commentStatus: Disabled
        id: id
        niceName: niceName
        email: email
        username: username
      properties:
        id:
          type: string
        username:
          type: string
        password:
          type: string
        niceName:
          type: string
        avatarUrl:
          type: string
        displayName:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        blogs:
          items:
            $ref: '#/components/schemas/Blog'
          type: array
          xml:
            name: blog
            wrapped: true
        verificationStatus:
          $ref: '#/components/schemas/VerificationStatus'
      type: object
      xml:
        name: User
    Role:
      example:
        capabilities:
        - capabilities
        - capabilities
        name: name
      properties:
        name:
          type: string
        capabilities:
          items:
            type: string
          type: array
      type: object
      xml:
        name: Role
    VerificationStatus:
      example:
        isVerified: true
        verificationMethod: verificationMethod
        verificationTimestamp: 2000-01-23T04:56:07.000+00:00
      properties:
        isVerified:
          type: boolean
        verificationTimestamp:
          format: date-time
          type: string
        verificationMethod:
          type: string
      type: object
      xml:
        name: VerificationStatus
    Blog:
      example:
        authToken: authToken
        name: name
        id: id
        url: url
        commentStatus: Disabled
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
        authToken:
          type: string
        commentStatus:
          enum:
          - Disabled
          - RequireReview
          - AutomateReview
          type: string
      type: object
      xml:
        name: Blog
...
    updatePostWithForm_request:
      properties:
        title:
          type: string
        url:
          type: string
        excerpt:
          type: string
        content:
          type: string
        commentStatus:
          enum:
          - Disabled
          - RequireReview
          - AutomateReview
          type: string
        postType:
          type: string
      type: object
...
  securitySchemes:
    cm_auth:
      flows:
        implicit:
          authorizationUrl: http://petstore.swagger.io/oauth/dialog
          scopes:
            write: modify
            read: read
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey