# servers/go/go/model_post.go
/*
 * Open APi
 *
 * This is a sample server. 
 *
 * API version: 1.0.0
 * Contact: email@gmail.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Post struct {

	Id string `json:"id,omitempty"`

	Title string `json:"title,omitempty"`

	Url string `json:"url,omitempty"`

	Excerpt string `json:"excerpt,omitempty"`

	Content string `json:"content,omitempty"`

	CommentStatus string `json:"commentStatus,omitempty"`

	PostType string `json:"postType,omitempty"`
}

// AssertPostRequired checks if the required fields are not zero-ed
func AssertPostRequired(obj Post) error {
	return nil
}

// AssertRecursePostRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of Post (e.g. [][]Post), otherwise ErrTypeAssertionError is thrown.
func AssertRecursePostRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aPost, ok := obj.(Post)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertPostRequired(aPost)
	})
}